pyproject.toml
[project]
name = "tpvalidator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alessandro Thea", email = "alessandro.thea@cern.ch" }
]
requires-python = ">=3.13"
dependencies = [
    "awkward>=2.8.3",
    "awkward-pandas>=2023.8.0",
    "beampy-slideshow>=0.5.5.post3",
    "click>=8.2.1",
    "cmocean>=4.0.3",
    "fpdf2>=2.8.4",
    "hist>=2.9.0",
    "ipykernel>=6.29.5",
    "ipympl>=0.9.7",
    "ipython>=9.2.0",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.3",
    "mpl-axes-aligner>=1.3",
    "mplhep>=0.4.0",
    "nbformat>=5.10.4",
    "pandas>=2.2.3",
    "plotly>=6.1.2",
    "reportlab>=4.4.3",
    "rich>=14.0.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "uproot>=5.6.2",
]

[project.scripts]
tpvalidator = "tpvalidator:main"
tp-basic-checks = "tpvalidator.basic:main"
tp-comparison = "tpvalidator.comparison:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
